Rename `CompilerBackend` to `CrossCompiler`
